combine search spec
